Move values_test under the tests package
