+++ runtime/interpreter/value_test.go
@@ -3140,463 +3140,3 @@ func TestNonStorable(t *testing.T) {
 	require.NoError(t, err)
 
 }
-
-func TestRandomMapOperations(t *testing.T) {
-	// TODO Skip by default
-
-	seed := time.Now().UnixNano()
-	fmt.Printf("Seed used for test: %d \n", seed)
-	rand.Seed(seed)
-
-	storage := NewInMemoryStorage()
-	inter, err := NewInterpreter(
-		nil,
-		utils.TestLocation,
-		WithStorage(storage),
-	)
-	require.NoError(t, err)
-
-	// TODO make this a param
-	numberOfValues := rand.Intn(100_000)
-
-	entries := make(map[interface{}]Value, numberOfValues)
-	var testMap, copyOfTestMap *DictionaryValue
-	var storageSize, slabCounts int
-	orgOwner := common.Address([8]byte{'A'})
-
-	t.Run("dictionary construction", func(t *testing.T) {
-		keyValues := make([]Value, numberOfValues*2)
-		for i := 0; i < numberOfValues; i++ {
-			// TODO maybe deep copy values
-
-			key := randomHashableValue(inter)
-			value := randomStorableValue(inter)
-
-			var mapKey interface{}
-
-			// Dereference string-keys before putting into go-map,
-			// as go-map hashing treats pointers as unique values.
-			// i.e: Maintain the value as the key, rather than the pointer.
-			switch key := deepCopyBasicValue(key).(type) {
-			case *StringValue:
-				mapKey = *key
-			case Value:
-				mapKey = key
-			default:
-				panic("unreachable")
-			}
-
-			entries[mapKey] = deepCopyBasicValue(value)
-
-			keyValues[i*2] = key
-			keyValues[i*2+1] = value
-		}
-
-		testMap = NewDictionaryValueWithAddress(inter,
-			DictionaryStaticType{
-				KeyType:   PrimitiveStaticTypeAnyStruct,
-				ValueType: PrimitiveStaticTypeAnyStruct,
-			},
-			orgOwner,
-			keyValues...,
-		)
-
-		storageSize, slabCounts = getSlabStorageSize(t, storage)
-
-		for orgKey, orgValue := range entries {
-			key := dictionaryKey(orgKey)
-
-			v, found := testMap.Get(inter, nil, key)
-			require.True(t, found)
-			require.Equal(t, v, orgValue, "key: %s", orgKey)
-
-			exists := testMap.ContainsKey(inter, nil, key)
-			require.True(t, bool(exists))
-		}
-		require.Equal(t, testMap.Count(), len(entries))
-		o := testMap.GetOwner()
-		require.Equal(t, o[:], orgOwner[:])
-	})
-
-	t.Run("test deep copy", func(t *testing.T) {
-		newOwner := atree.Address([8]byte{'B'})
-		copyOfTestMap = testMap.DeepCopy(inter, nil, newOwner).(*DictionaryValue)
-
-		for orgKey, orgValue := range entries {
-			key := dictionaryKey(orgKey)
-
-			v, found := copyOfTestMap.Get(inter, nil, key)
-			require.True(t, found)
-			require.Equal(t, v, orgValue)
-
-			exists := copyOfTestMap.ContainsKey(inter, nil, key)
-			require.True(t, bool(exists))
-		}
-		require.Equal(t, copyOfTestMap.Count(), len(entries))
-		o := copyOfTestMap.GetOwner()
-		require.Equal(t, o[:], newOwner[:])
-	})
-
-	t.Run("test deep removal", func(t *testing.T) {
-		copyOfTestMap.DeepRemove(inter)
-		err = storage.Remove(copyOfTestMap.StorageID())
-		require.NoError(t, err)
-
-		// deep removal should clean up everything
-		newStorageSize, newSlabCounts := getSlabStorageSize(t, storage)
-		require.Equal(t, slabCounts, newSlabCounts)
-		require.Equal(t, storageSize, newStorageSize)
-
-		// go over original values again and check no missing data (no side effect should be found)
-		for orgKey, orgValue := range entries {
-			key := dictionaryKey(orgKey)
-
-			v, found := testMap.Get(inter, nil, key)
-			require.True(t, found)
-			require.Equal(t, v, orgValue)
-
-			exists := testMap.ContainsKey(inter, nil, key)
-			require.True(t, bool(exists))
-		}
-
-		require.Equal(t, testMap.Count(), len(entries))
-		o := testMap.GetOwner()
-		require.Equal(t, o[:], orgOwner[:])
-	})
-
-	t.Run("test iterator", func(t *testing.T) {
-		// TODO
-	})
-
-	// temp to prevent test cache
-	// t.Fatal("XXX")
-
-}
-
-func dictionaryKey(i interface{}) Value {
-	switch key := i.(type) {
-	case StringValue:
-		return &key
-	case Value:
-		return key
-	default:
-		panic("unreachable")
-	}
-}
-
-func getSlabStorageSize(t *testing.T, storage InMemoryStorage) (totalSize int, slabCounts int) {
-	slabs, err := storage.Encode()
-	require.NoError(t, err)
-	for _, slab := range slabs {
-		totalSize += len(slab)
-		slabCounts++
-	}
-	return
-}
-
-func deepCopyBasicValue(value Value) Value {
-	switch v := value.(type) {
-
-	// Int
-	case IntValue:
-		var n big.Int
-		n.Set(v.BigInt)
-		return NewIntValueFromBigInt(&n)
-	case Int8Value:
-		return Int8Value(int8(v))
-	case Int16Value:
-		return Int16Value(int16(v))
-	case Int32Value:
-		return Int32Value(int32(v))
-	case Int64Value:
-		return Int64Value(int64(v))
-	case Int128Value:
-		var n big.Int
-		n.Set(v.BigInt)
-		return NewInt128ValueFromBigInt(&n)
-	case Int256Value:
-		var n big.Int
-		n.Set(v.BigInt)
-		return NewInt256ValueFromBigInt(&n)
-
-	// Uint
-	case UIntValue:
-		var n big.Int
-		n.Set(v.BigInt)
-		return NewUIntValueFromBigInt(&n)
-	case UInt8Value:
-		return UInt8Value(uint8(v))
-	case UInt16Value:
-		return UInt16Value(uint16(v))
-	case UInt32Value:
-		return UInt32Value(uint32(v))
-	case UInt64Value:
-		return UInt64Value(uint64(v))
-	case UInt128Value:
-		var n big.Int
-		n.Set(v.BigInt)
-		return NewUInt128ValueFromBigInt(&n)
-	case UInt256Value:
-		var n big.Int
-		n.Set(v.BigInt)
-		return NewUInt256ValueFromBigInt(&n)
-
-	case Word8Value, Word16Value, Word32Value, Word64Value:
-		return v
-
-	case *StringValue:
-		b := []byte(v.Str)
-		data := make([]byte, len(b))
-		copy(data, b)
-		return NewStringValue(string(data))
-
-	case AddressValue:
-		b := v[:]
-		data := make([]byte, len(b))
-		copy(data, b)
-		return NewAddressValueFromBytes(data)
-	case Fix64Value:
-		return NewFix64ValueWithInteger(int64(v.ToInt()))
-	case UFix64Value:
-		return NewUFix64ValueWithInteger(uint64(v.ToInt()))
-
-	case PathValue:
-		return PathValue{
-			Domain:     v.Domain,
-			Identifier: v.Identifier,
-		}
-
-	case BoolValue:
-		return v
-
-	default:
-		return NilValue{}
-	}
-}
-
-func randomStorableValue(interpreter *Interpreter) Value {
-	n := rand.Intn(Enum)
-	switch n {
-	case Void:
-		return VoidValue{}
-	case Nil:
-		return NilValue{}
-	default:
-		return generateRandomHashableValue(interpreter, n)
-	}
-}
-
-func randomHashableValue(interpreter *Interpreter) Value {
-	return generateRandomHashableValue(interpreter, rand.Intn(Enum))
-}
-
-func generateRandomHashableValue(interpreter *Interpreter, n int) Value {
-	switch n {
-	// TODO deal with negative numbers
-
-	// Int
-	case Int:
-		return NewIntValueFromInt64(rand.Int63())
-	case Int8:
-		return Int8Value(rand.Intn(255))
-	case Int16:
-		return Int16Value(rand.Intn(65535))
-	case Int32:
-		return Int32Value(rand.Int31())
-	case Int64:
-		return Int64Value(rand.Int63())
-	case Int128:
-		return NewInt128ValueFromInt64(rand.Int63())
-	case Int256:
-		return NewInt256ValueFromInt64(rand.Int63())
-
-	// UInt
-	case UInt:
-		return NewUIntValueFromUint64(rand.Uint64())
-	case UInt8:
-		return UInt8Value(rand.Intn(255))
-	case UInt16:
-		return UInt16Value(rand.Intn(65535))
-	case UInt32:
-		return UInt32Value(rand.Uint32())
-	case UInt64_1, UInt64_2, UInt64_3, UInt64_4: // should be more common
-		return UInt64Value(rand.Uint64())
-	case UInt128:
-		return NewUInt128ValueFromBigInt(big.NewInt(rand.Int63()))
-	case UInt256:
-		return NewUInt256ValueFromBigInt(big.NewInt(rand.Int63()))
-
-	// Word
-	case Word8:
-		return Word8Value(rand.Intn(255))
-	case Word16:
-		return Word16Value(rand.Intn(65535))
-	case Word32:
-		return Word32Value(rand.Uint32())
-	case Word64:
-		return Word64Value(rand.Uint64())
-
-	// Fixed point
-	case Fix64:
-		return NewFix64ValueWithInteger(rand.Int63n(sema.Fix64TypeMaxInt))
-	case UFix64:
-		return NewUFix64ValueWithInteger(
-			uint64(rand.Int63n(
-				int64(sema.UFix64TypeMaxInt),
-			)),
-		)
-
-	// String
-	case String_1, String_2, String_3, String_4: // small string - should be more common
-		size := rand.Intn(255)
-		data := make([]byte, size)
-		rand.Read(data)
-		return NewStringValue(string(data))
-	case String_5: // large string
-		size := rand.Intn(4048) + 255
-		data := make([]byte, size)
-		rand.Read(data)
-		return NewStringValue(string(data))
-
-	case Bool_True:
-		return BoolValue(true)
-	case Bool_False:
-		return BoolValue(false)
-
-	case Address:
-		data := make([]byte, 8)
-		rand.Read(data)
-		return NewAddressValueFromBytes(data)
-
-	case Path:
-		randomDomain := rand.Intn(len(common.AllPathDomains))
-		identifier := make([]byte, 8)
-		rand.Read(identifier)
-
-		return PathValue{
-			Domain:     common.AllPathDomains[randomDomain],
-			Identifier: string(identifier),
-		}
-
-	case Enum:
-		typ := rand.Intn(Word64)
-		rawValue := generateRandomHashableValue(interpreter, typ).(NumberValue)
-
-		identifier := make([]byte, 8)
-		rand.Read(identifier)
-
-		return NewEnumCaseValue(
-			interpreter,
-			&sema.CompositeType{
-				Identifier:  string(identifier),
-				EnumRawType: intSubtype(typ),
-				Kind:        common.CompositeKindEnum,
-			},
-			rawValue,
-			nil,
-		)
-
-	default:
-		panic(fmt.Sprintf("unsupported:  %d", n))
-	}
-}
-
-func intSubtype(n int) sema.Type {
-	switch n {
-	// Int
-	case Int:
-		return sema.IntType
-	case Int8:
-		return sema.Int8Type
-	case Int16:
-		return sema.Int16Type
-	case Int32:
-		return sema.Int32Type
-	case Int64:
-		return sema.Int64Type
-	case Int128:
-		return sema.Int128Type
-	case Int256:
-		return sema.Int256Type
-
-	// UInt
-	case UInt:
-		return sema.UIntType
-	case UInt8:
-		return sema.UInt8Type
-	case UInt16:
-		return sema.UInt16Type
-	case UInt32:
-		return sema.UInt32Type
-	case UInt64_1, UInt64_2, UInt64_3, UInt64_4:
-		return sema.UInt64Type
-	case UInt128:
-		return sema.UInt128Type
-	case UInt256:
-		return sema.UInt256Type
-
-	// Word
-	case Word8:
-		return sema.Word8Type
-	case Word16:
-		return sema.Word16Type
-	case Word32:
-		return sema.Word32Type
-	case Word64:
-		return sema.Word64Type
-
-	default:
-		panic(fmt.Sprintf("unsupported:  %d", n))
-	}
-}
-
-const (
-	// Hashable values
-	Int = iota
-	Int8
-	Int16
-	Int32
-	Int64
-	Int128
-	Int256
-
-	UInt
-	UInt8
-	UInt16
-	UInt32
-	UInt64_1
-	UInt64_2
-	UInt64_3
-	UInt64_4
-	UInt128
-	UInt256
-
-	Word8
-	Word16
-	Word32
-	Word64
-
-	Fix64
-	UFix64
-
-	String_1
-	String_2
-	String_3
-	String_4
-	String_5
-
-	Bool_True
-	Bool_False
-	Path
-	Address
-	Enum
-
-	// Non-hashable values
-
-	Void
-	Nil // `Never?`
-
-	Struct
-	Resource
-	Contract
-)
